# Implement Liquid::C::Expression to optimize expression evaluation