--- a/ext/liquid_c/c_buffer.h
+++ b/ext/liquid_c/c_buffer.h
@@ -54,4 +54,9 @@ inline void c_buffer_rb_gc_mark(c_buffer_t *buffer)
     }
 }
 
+inline void c_buffer_concat(c_buffer_t *dest, c_buffer_t *src)
+{
+    c_buffer_write(dest, src->data, c_buffer_size(src));
+}
+
 #endif