--- a/ext/liquid_c/parser.c
+++ b/ext/liquid_c/parser.c
@@ -2,8 +2,8 @@
 #include "parser.h"
 #include "lexer.h"
 
-static VALUE cLiquidRangeLookup, cLiquidVariableLookup, cRange, vLiquidExpressionLiterals;
-static ID idToI, idEvaluate;
+static VALUE empty_string;
+static ID id_to_i, idEvaluate;
 
 void init_parser(parser_t *p, const char *str, const char *end)
 {
@@ -67,92 +67,133 @@ static VALUE parse_number(parser_t *p)
     return out;
 }
 
-static VALUE parse_range(parser_t *p)
+static VALUE try_parse_constant_range(parser_t *p)
 {
+    parser_t saved_state = *p;
+
     parser_must_consume(p, TOKEN_OPEN_ROUND);
 
-    VALUE args[2];
-    args[0] = parse_expression(p);
+    VALUE begin = try_parse_constant_expression(p);
+    if (begin == Qundef) {
+        *p = saved_state;
+        return Qundef;
+    }
     parser_must_consume(p, TOKEN_DOTDOT);
 
-    args[1] = parse_expression(p);
+    VALUE end = try_parse_constant_expression(p);
+    if (end == Qundef) {
+        *p = saved_state;
+        return Qundef;
+    }
     parser_must_consume(p, TOKEN_CLOSE_ROUND);
 
-    if (rb_respond_to(args[0], idEvaluate) || rb_respond_to(args[1], idEvaluate))
-        return rb_class_new_instance(2, args, cLiquidRangeLookup);
+    begin = rb_funcall(begin, id_to_i, 0);
+    end = rb_funcall(end, id_to_i, 0);
 
-    return rb_class_new_instance(2, args, cRange);
+    bool exclude_end = false;
+    return rb_range_new(begin, end, exclude_end);
 }
 
-static VALUE parse_variable(parser_t *p)
+static void parse_and_compile_range(parser_t *p, vm_assembler_t *code)
 {
-    VALUE name, lookups = rb_ary_new(), lookup;
-    unsigned long long command_flags = 0;
-    bool static_variable_name = false;
+    VALUE const_range = try_parse_constant_range(p);
+    if (const_range != Qundef) {
+        vm_assembler_add_push_const(code, const_range);
+        return;
+    }
 
+    parser_must_consume(p, TOKEN_OPEN_ROUND);
+    parse_and_compile_expression(p, code);
+    parser_must_consume(p, TOKEN_DOTDOT);
+    parse_and_compile_expression(p, code);
+    parser_must_consume(p, TOKEN_CLOSE_ROUND);
+    vm_assembler_add_new_int_range(code);
+}
+
+static void parse_and_compile_variable_lookup(parser_t *p, vm_assembler_t *code)
+{
     if (parser_consume(p, TOKEN_OPEN_SQUARE).type) {
-        name = parse_expression(p);
+        parse_and_compile_expression(p, code);
         parser_must_consume(p, TOKEN_CLOSE_SQUARE);
+        vm_assembler_add_find_variable(code);
     } else {
-        static_variable_name = true;
-        name = token_to_rstr(parser_must_consume(p, TOKEN_IDENTIFIER));
+        VALUE name = token_to_rstr_leveraging_existing_symbol(parser_must_consume(p, TOKEN_IDENTIFIER));
+        vm_assembler_add_find_static_variable(code, name);
     }
 
     while (true) {
         if (p->cur.type == TOKEN_OPEN_SQUARE) {
             parser_consume_any(p);
-            lookup = parse_expression(p);
+            parse_and_compile_expression(p, code);
             parser_must_consume(p, TOKEN_CLOSE_SQUARE);
-
-            rb_ary_push(lookups, lookup);
+            vm_assembler_add_lookup_key(code);
         } else if (p->cur.type == TOKEN_DOT) {
             int has_space_affix = parser_consume_any(p).flags & TOKEN_SPACE_AFFIX;
-            lookup = token_to_rstr(parser_must_consume(p, TOKEN_IDENTIFIER));
+            VALUE key = token_to_rstr_leveraging_existing_symbol(parser_must_consume(p, TOKEN_IDENTIFIER));
 
             if (has_space_affix)
                 rb_enc_raise(utf8_encoding, cLiquidSyntaxError, "Unexpected dot");
 
-            if (rstring_eq(lookup, "size") || rstring_eq(lookup, "first") || rstring_eq(lookup, "last"))
-                command_flags |= 1 << RARRAY_LEN(lookups);
-
-            rb_ary_push(lookups, lookup);
+            if (rstring_eq(key, "size") || rstring_eq(key, "first") || rstring_eq(key, "last"))
+                vm_assembler_add_lookup_command(code, key);
+            else
+                vm_assembler_add_lookup_const_key(code, key);
         } else {
             break;
         }
     }
-
-    if (RARRAY_LEN(lookups) == 0 && static_variable_name) {
-        VALUE literal = rb_hash_lookup2(vLiquidExpressionLiterals, name, Qundef);
-        if (literal != Qundef) return literal;
-    }
-
-    VALUE args[4] = {Qfalse, name, lookups, INT2FIX(command_flags)};
-    return rb_class_new_instance(4, args, cLiquidVariableLookup);
 }
 
-bool will_parse_constant_expression_next(parser_t *p)
+static VALUE try_parse_literal(parser_t *p)
+{
+    const char *str = p->cur.val;
+    long size = p->cur.val_end - str;
+    VALUE result = Qundef;
+    switch (size) {
+        case 3:
+            if (memcmp(str, "nil", size) == 0)
+                result = Qnil;
+            break;
+        case 4:
+            if (memcmp(str, "null", size) == 0) {
+                result = Qnil;
+            } else if (memcmp(str, "true", size) == 0) {
+                result = Qtrue;
+            }
+            break;
+        case 5:
+            switch (*str) {
+                case 'f':
+                    if (memcmp(str, "false", size) == 0)
+                        result = Qfalse;
+                    break;
+                case 'b':
+                    if (memcmp(str, "blank", size) == 0)
+                        result = empty_string;
+                    break;
+                case 'e':
+                    if (memcmp(str, "empty", size) == 0)
+                        result = empty_string;
+                    break;
+            }
+            break;
+    }
+    if (result != Qundef)
+        parser_consume_any(p);
+    return result;
+}
+
+VALUE try_parse_constant_expression(parser_t *p)
 {
     switch (p->cur.type) {
         case TOKEN_IDENTIFIER:
-        case TOKEN_OPEN_SQUARE:
-            return false;
-        default:
-            return true;
-    }
-}
-
-VALUE parse_expression(parser_t *p)
-{
-    switch (p->cur.type) {
-        case TOKEN_IDENTIFIER:
-        case TOKEN_OPEN_SQUARE:
-            return parse_variable(p);
+            return try_parse_literal(p);
 
         case TOKEN_NUMBER:
             return parse_number(p);
 
         case TOKEN_OPEN_ROUND:
-            return parse_range(p);
+            return try_parse_constant_range(p);
 
         case TOKEN_STRING:
         {
@@ -162,6 +203,53 @@ VALUE parse_expression(parser_t *p)
             return token_to_rstr(token);
         }
     }
+    return Qundef;
+}
+
+static void parse_and_compile_number(parser_t *p, vm_assembler_t *code)
+{
+    VALUE num = parse_number(p);
+    if (RB_FIXNUM_P(num))
+        vm_assembler_add_push_fixnum(code, num);
+    else
+        vm_assembler_add_push_const(code, num);
+    return;
+}
+
+void parse_and_compile_expression(parser_t *p, vm_assembler_t *code)
+{
+    switch (p->cur.type) {
+        case TOKEN_IDENTIFIER:
+        {
+            VALUE literal = try_parse_literal(p);
+            if (literal != Qundef) {
+                vm_assembler_add_push_literal(code, literal);
+                return;
+            }
+            // fallthrough
+        }
+        case TOKEN_OPEN_SQUARE:
+            parse_and_compile_variable_lookup(p, code);
+            return;
+
+        case TOKEN_NUMBER:
+            parse_and_compile_number(p, code);
+            return;
+
+        case TOKEN_OPEN_ROUND:
+            parse_and_compile_range(p, code);
+            return;
+
+        case TOKEN_STRING:
+        {
+            lexer_token_t token = parser_consume_any(p);
+            token.val++;
+            token.val_end--;
+            VALUE str = token_to_rstr(token);
+            vm_assembler_add_push_const(code, str);
+            return;
+        }
+    }
 
     if (p->cur.type == TOKEN_EOS) {
         rb_enc_raise(utf8_encoding, cLiquidSyntaxError, "[:%s] is not a valid expression", symbol_names[p->cur.type]);
@@ -169,46 +257,14 @@ VALUE parse_expression(parser_t *p)
         rb_enc_raise(utf8_encoding, cLiquidSyntaxError, "[:%s, \"%.*s\"] is not a valid expression",
                  symbol_names[p->cur.type], (int)(p->cur.val_end - p->cur.val), p->cur.val);
     }
-    return Qnil;
-}
-
-static VALUE rb_parse_expression(VALUE self, VALUE markup)
-{
-    StringValue(markup);
-    char *start = RSTRING_PTR(markup);
-
-    parser_t p;
-    init_parser(&p, start, start + RSTRING_LEN(markup));
-
-    if (p.cur.type == TOKEN_EOS)
-        return Qnil;
-
-    VALUE expr = parse_expression(&p);
-
-    if (p.cur.type != TOKEN_EOS)
-        rb_enc_raise(utf8_encoding, cLiquidSyntaxError, "[:%s] is not a valid expression", symbol_names[p.cur.type]);
-
-    return expr;
 }
 
 void init_liquid_parser(void)
 {
-    idToI = rb_intern("to_i");
+    id_to_i = rb_intern("to_i");
     idEvaluate = rb_intern("evaluate");
 
-    cLiquidRangeLookup = rb_const_get(mLiquid, rb_intern("RangeLookup"));
-    rb_global_variable(&cLiquidRangeLookup);
-
-    cRange = rb_const_get(rb_cObject, rb_intern("Range"));
-    rb_global_variable(&cRange);
-
-    cLiquidVariableLookup = rb_const_get(mLiquid, rb_intern("VariableLookup"));
-    rb_global_variable(&cLiquidVariableLookup);
-
-    VALUE cLiquidExpression = rb_const_get(mLiquid, rb_intern("Expression"));
-    rb_define_singleton_method(cLiquidExpression, "c_parse", rb_parse_expression, 1);
-
-    vLiquidExpressionLiterals = rb_const_get(cLiquidExpression, rb_intern("LITERALS"));
-    rb_global_variable(&vLiquidExpressionLiterals);
+    empty_string = rb_utf8_str_new_literal("");
+    rb_global_variable(&empty_string);
 }