--- a/ext/liquid_c/variable_lookup.h
+++ b/ext/liquid_c/variable_lookup.h
@@ -2,7 +2,7 @@
 #define LIQUID_VARIABLE_LOOKUP_H
 
 void init_liquid_variable_lookup();
-VALUE variable_lookup_evaluate(VALUE self, VALUE expression);
+VALUE variable_lookup_key(VALUE context, VALUE object, VALUE key, bool is_command);
 
 #endif