--- a/ext/liquid_c/vm.h
+++ b/ext/liquid_c/vm.h
@@ -8,5 +8,6 @@ void init_liquid_vm();
 void liquid_vm_render(block_body_t *block, VALUE context, VALUE output);
 void liquid_vm_next_instruction(const uint8_t **ip_ptr, const size_t **const_ptr_ptr);
 bool liquid_vm_filtering(VALUE context);
+VALUE liquid_vm_evaluate(VALUE context, vm_assembler_t *code);
 
 #endif