… ext/liquid_c/vm.c
… ext/liquid_c/vm_assembler.c
… ext/liquid_c/vm_assembler.h
… lib/liquid/c.rb
… test/liquid_test.rb
… test/unit/context_test.rb
… test/unit/expression_test.rb
… test/unit/variable_test.rb