~~Depends on https://github.com/Shopify/liquid-c/pull/59 and https://github.com/Shopify/liquid/pull/1300~~

## Problem

In https://github.com/Shopify/liquid-c/pull/59, expressions were compiled to ruby constants or lookup ruby objects (Liquid::VariableLookup & Liquid::RangeLookup) that are evaluated with `context_evaluate`.  This was done in that PR for simplicity, but they should really be compiled to more granular VM instructions for rendering performance and for serializability (for future caching of VM compiled templates).

## Solution

I replaced the previous code for strict parsing expressions and variable lookups to instead emit VM code, which is now compiled directly into Liquid::C::BlockBody objects when compiling variables.  Expressions are also often parsed and evaluated independent of Liquid::Variable in tags, so I introduced a `Liquid::C::Expression` object that gets returned from `Liquid::Expression.parse` if the expression includes a variable lookup and strict parsing of it succeeds.  I ended up doing both of these in the same commit, since I also leverage the `Liquid::C::Expression` object for filter keyword argument parsing to handle automatic cleanup on a parse exception.

Note that a couple of direct uses of `Liquid::Variable.new` (in assign and echo tags) no longer leverage liquid-c for parsing, which slightly impacts parse speed.  I'll embed those tags in the liquid VM code in a follow-up PR rather than introducing a Liquid::C::Variable class.  This PR significantly speeds up rendering, as shown in the benchmarks, which more than makes up for this parsing regression.

Since `Liquid::Variable.new` is no longer monkey patched and `Liquid::Expression.parse` no longer turns a comparable object, I had to update the liquid-c variable unit tests so they are tested through normal template parsing and rendering like an integration test.  I also stopped testing against the liquid gem's variable unit tests for the same reasons.

To help with detecting and debugging VM stack underflows or overflows, I added some assertions that are enabled by default for CI and gem development but are disabled by default for gem installation.

## Benchmarks

Before this PR (on the https://github.com/Shopify/liquid-c/pull/59 branch)

```
              parse:    167.388  (± 2.4%) i/s -      1.680k in  10.043356s
             render:    173.196  (± 3.5%) i/s -      1.734k in  10.022222s
     parse & render:     80.409  (± 1.2%) i/s -    808.000  in  10.050380s
```

after

```
              parse:    162.660  (± 2.5%) i/s -      1.632k in  10.040800s
             render:    231.388  (± 4.3%) i/s -      2.323k in  10.059088s
     parse & render:     86.735  (± 1.2%) i/s -    872.000  in  10.056157s
```